tempfix #6186 + minor the hard deprecation of gl.LineWidth caused stack overflows due to widgets crashing between Push and Pop

+++ rts/Rendering/LineDrawer.cpp
@@ -108,7 +108,7 @@ void CLineDrawer::DrawAll(bool onMiniMap)
 	shader->Disable();
 	glAttribStatePtr->PopBits();
 
-	for (int i = 0; i < regularLines.size(); i++) {
+	for (size_t i = 0; i < regularLines.size(); i++) {
 		regularLines[i].clear();
 		stippleLines[i].clear();
 	}
